requirements.txt
certifi==2018.11.29
chardet==3.0.4
codecov==2.0.15
coverage==4.5.2
Django==2.1.3
django-fsm==2.6.0
djangorestframework==3.9.0
idna==2.8
pytz==2018.7
requests==2.21.0
urllib3==1.24.1
